Revert "Pass time to RgbFunctionFactory (#10)"
This reverts commit 0492ab9abe1bdebdd26abab83869b9774c76c0aa.

+++ lib/Processing/Interfaces/ProcessingTypes.h
@@ -117,7 +117,7 @@ Json convert(const TNoteToLightMap& source);
 TNoteToLightMap convert(const Json& source);
 
 /** Type for actual time in milliseconds. */
-typedef uint32_t TTime;
+typedef uint64_t TTime;
 
 /** Type for actual note states. */
 struct TNoteState
